upload: show error message if set from server
If the server has set an error message and the error isn't one we
already have explicit error handling for, we default to showing it.

This is useful as I've just added a rate limiting error in the backend,
and the error should be presented as is to the user.