Refactor the DOXA CLI
The handling of configuration has been refactored. Specifying an environment is now optional. There is a new mechanism to allow the server to encourage clients to update to the latest version if their installed version of the CLI is outdated. Furthermore, `doxa.yaml` files have been made optional if a competition is specified in the command (which may be useful in the future).